+++ src/TableCaptionRenderer.php
@@ -1,0 +1,39 @@
+<?php
+
+/*
+ * This is part of the webuni/commonmark-table-extension package.
+ *
+ * (c) Martin Hasoň <martin.hason@gmail.com>
+ * (c) Webuni s.r.o. <info@webuni.cz>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Webuni\CommonMark\TableExtension;
+
+use League\CommonMark\Block\Element\AbstractBlock;
+use League\CommonMark\Block\Renderer\BlockRendererInterface;
+use League\CommonMark\ElementRendererInterface;
+use League\CommonMark\HtmlElement;
+
+class TableCaptionRenderer implements BlockRendererInterface
+{
+    public function render(AbstractBlock $block, ElementRendererInterface $htmlRenderer, $inTightList = false)
+    {
+        if (!($block instanceof TableCaption)) {
+            throw new \InvalidArgumentException('Incompatible block type: '.get_class($block));
+        }
+
+        $attrs = [];
+        foreach ($block->getData('attributes', []) as $key => $value) {
+            $attrs[$key] = $htmlRenderer->escape($value, true);
+        }
+
+        if ($block->id) {
+            $attrs['id'] = $block->id;
+        }
+
+        return new HtmlElement('caption', $attrs, $htmlRenderer->renderInlines($block->children()));
+    }
+}